refactor: address review — styles in component, prop types inline
- Move styles from styles.ts into XDSCalendar.tsx (bottom of file)
- Keep prop types in the component file for easy introspection
- types.ts now only holds shared re-exports (composition types)
- Preserve SYNC comments
- Update CODE_STYLE.md to reflect: internal vs public hooks distinction,
  prop types stay in component file, styles at bottom position 5
